# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Syndikat Disc Golf API
  version: 1.0.0
  description: |
    Website-facing API for syndikat.golf. This contract is maintained manually alongside Express routes.
    Third-party scraper data can change independently; clients must tolerate omitted optional fields.
  license:
    name: ISC
    identifier: ISC
  contact:
    name: Syndikat Disc Golf
    url: https://syndikat.golf
servers:
  - url: https://api.syndikat.golf
    description: Production
  - url: http://{host}:8080
    description: Local development
    variables:
      host:
        default: localhost
# All endpoints are public unless an operation overrides this with a security requirement.
security: []
...
paths:
...
  /training/status:
    get:
      tags: [Training]
      summary: Return the next Friday training date and whether signup is open
      description: Returns the next training date and its current signup availability.
      operationId: getTrainingStatus
      responses:
        '200':
          description: Signup status, calculated in the Europe/Berlin timezone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingStatus'
        '503':
          $ref: '#/components/responses/TrainingUnavailable'
  /training/participants:
    get:
      tags: [Training]
      summary: List participants for a training date
      description: Returns participants registered for the requested training date.
      operationId: listTrainingParticipants
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Participants ordered by signup time
          content:
            application/json:
              schema:
                type: object
                required: [participants]
                properties:
                  participants:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrainingParticipant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/TrainingUnavailable'
    post:
      tags: [Training]
      summary: Sign up for the next Friday training
      description: Requires the current signup password. Save the returned removal token locally; it is not retrievable later.
      operationId: createTrainingParticipant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, password]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 40
                password:
                  type: string
                  writeOnly: true
      responses:
        '201':
          description: Participant created
          content:
            application/json:
              schema:
                type: object
                required: [participant, removalToken]
                properties:
                  participant:
                    $ref: '#/components/schemas/CreatedTrainingParticipant'
                  removalToken:
                    type: string
                    description: Secret bearer token used only to remove this participant.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Incorrect signup password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Signup is closed after Friday 18:00 Europe/Berlin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '409':
          description: The name is already signed up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/TrainingUnavailable'
  /training/participants/{id}:
    delete:
      tags: [Training]
      summary: Remove a participant
      description: Removes a participant using the removal token returned at signup.
      operationId: deleteTrainingParticipant
      security:
        - RemovalToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Participant removed
        '401':
          description: Missing, invalid, or non-matching removal token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/TrainingUnavailable'
components:
  securitySchemes:
    RemovalToken:
      type: http
      scheme: bearer
      bearerFormat: opaque removal token
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    ServerError:
      description: Server or upstream error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    TrainingUnavailable:
      description: Training signup has not been configured
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
  schemas:
...
    Message:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    TrainingStatus:
      type: object
      required: [date, signupOpen]
      properties:
        date: { type: string, format: date }
        signupOpen: { type: boolean }
    TrainingParticipant:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    CreatedTrainingParticipant:
      allOf:
        - $ref: '#/components/schemas/TrainingParticipant'
        - type: object
          required: [date]
          properties:
            date: { type: string, format: date }